Add support for PackBeam to take AVM files as input.
This change set modifies the PackBeam executable to support previously created AVM files as inputs.  This allows the creation of AVM "archives", or "libraries", which can be built independently of an application (e.g., estdlib), and which can then be assembled into an AVM that can be flashed to a device or run from the command line.

These changes are backwards compatible with previous uses of PackBeam.

These changes are submitted under the terms of the LGPLv2 and Apache2 licenses.