Adds support for building Erlang libraries and examples, and fixes a bug with PackBEAM.
This change set adds CMake rules for building AVM files for libraries and examples.

In addition, a bug in the PackBEAM executable is fixed, which fixes an issue with packing previously packed AVM files.

As part of this fix, the avmpack_fold fold function signature was changed, to supply accurate section addresses.  The API documentation was clarified and corrected.

These changes are submitted under the terms of the LGPLv2 and Apache2 licenses.