Load driver from system-dependent library.
This change reverts a change that was made in the initial socket driver changeset, which had moved the initialization of drivers to the libAtomVM shlib.  This change reverts this behavior, so that the driver is loaded from the platform-dependent library.  The advantage to loading the drivers this way is that other ports are not obligated to implement all supported drivers, or can add their own port drivers without affecting the libAtomVM core.

An attempt is still made to consolidate common functionality, such as term input/output handling between the driver and the application, in the libAtomVM.  A circular dependency between the platform library and platform dependent library is defined for the UNIX port.  The previous udpdriver modules were removed, as theire functionality has been included in the generic socket driver.

These changes are submitted under the terms of the LGPLv2 and Apache2 licenses.