Add selective receive to the console port.
This change adds selective receive when sending messages to the console port, using a reference to tag the request.  This way, the calling application can be assured that the request it made was serviced by the port.

This change also adds a flush operation to the console port, to flush any messages buffered by the standard library or OS.

Corresponding changes are made to the console erlang application.

These changes are made under the terms of the LGPLv2 and Apache2 licenses.