Ensure adequate memory on UDP receive.
This change ensures there is adequate room in the process heap to store enough data received on a UDP packet.

This change imposes an artificial limitation of 128 bytes on a received packet.  We will revisit this limitation in the future.

This change also fixes a potential memory corruption issue when copying an erlang reference to the event listener for the receive operation.

These changes are made under the terms of the LGPLv2 and Apache2 licenses.