Phase I changes to the network port driver.
This change set makes an initial set of improvements to the network port driver, including:

* Factored out common parts of the port driver code (the parts that interface directly with AtomVM i/o parameters) into a common set of functions in the libAtomVM module, similar to patterns used in the socket and console drivers;
* Modified the protocol between AtomVM and the network driver to use a reference, to support selective receive;
* Added a setup_network example.

This is part of a multi-phase change to this driver.  Subsequent changes will include a re-design of the network interface, to allow applications to respond to events in the Wifi driver, such as connect, disconnect, and IP acquisition events, as well as acquisition of netwrk information (e.g., ifconfig) and documentation as the interfaces settle down.  Further changes after that will support SoftAP and combined SoftAP+STA mode, as well as support for wifi scan.

These changes are made under the terms of the LGPLv2 and Apache2 licenses.