Add support for generating documentation.
This change set adds support for generating documentation for AtomVM.

If doxygen is available on the users path (located via CMake), then the
doc target may be used to generate documentation under build/doc/html.
Documentation is built for C sources, Erlang sources, and from markdown
source documents in the doc directory.  The doc target must be called
explicitly to generate documentation, and little or no dependency
checking is done to circumvent rebuilds.

Not all APIs and interfaces are currently documented.  These changes just
add infrastructure and some documentation to get started.

These changes are submitted under the terms of the LGPLv2 and Apache2
licenses.