Add support for apply/1 (112) opcode.
This PR adds support for the apply/1 opcode (#112), which supports Module:Function(...) calls.

An additional common function is added to the global_context module to simplify the retrieval of AtomStrings from term objects, and some code pths were cleaned up to allow for the possibility of loading code that does not exist (user error).  A unit test is added for exported module and NIF use cases, as well as for error conditions.

These changes are made under the terms of the LGPLv2 and Apache2 licenses.