Added script for cleaning whitespace in source modules.
This commit adds a script for cleaning source modules so that there are no modules with trailing whitespace and so that files end in the newline character appropriate for the platform.

In addition, whitespace fixes to existing source modules are made.

These changes are made under the terms of the LGPLv2 and Apache2 licenses.