Add support for testing Erlang libraries.
This change set adds CMake rules and a framework for testing Erlang libraries that are built as part of AtomVM.  A few small sample tests are added, which can be enhanced in the future.

In addition, this change set makes a few small changes to the AtomVM source code to better support testing:

* Improved error handling and reporting in cases where modules and/or functions are not resolvable
* Factored out common code for reporting "module:function/arity" for the above change
* Added the erlang:timestamp NIF for testing some timing functionality
* Increased the heap size when spawning a function to avoid memory corruption after gc events
* Added a unit test for the erlang:timestamp NIF
* Changed the AtomVM executable to return 0 if the return from the start/0 entrypoint is the atom "ok", or 0, otherwise
* Added CMake rules and an etest library for testing erlang libraries
* Added sample tests for estdlib.
* Turns off Erlang builds when the Xcode generator is enabled (reduces noise in Xcode projects)

These changes are made under the terms of the LGPLv2 and Apache2 licenses.