Change the names of modules in the estdlib library.
This change set changes the names of the estdlib library to use names that are not the same
as their OTP equivalents.  All modules in this library are prefixed with 'avm_'.  For
convenience, macros are provided that can be used to insulate the user from these names.

In addition, a few bug fixes and enhancements were made to the logger and avm_lists modules.

This change is made under the terms of the LGPLv2 and Apache2 licenses.