Fix support for funs in AVM files
This change set fixes support for funs in AVM (packed beam) files.  In addition, it fixes some minor issues with calling funs defined in other modules, and add functions and tests in the avm_lists module for using funs.

These changes are made under the terms of the LGPLv2 and Apache2 licenses.