Fixed OP_PUT_LIST opcode
This change set moves the allocation of a list element in the OP_PUT_LIST handling code after the decoding of terms, which may trigger the garbage collector, and could otherwise invalidate the allocated list element in the heap.

These changes are made under the terms of the LGPLv2 and Apache2 licenses.