Fixed calculation of boxed size of binaries.
The calculation of the boxed size of a binary is off by 1, which causes the binary to not be properly (shallow) copied during GC and in message passing.

Also modified socket to use binaries and calculate correct heap size.

This fix is made under the terms of the LGPLv2 and Apache2 licenses.