Enhanced Network API
This change set enhances the network API to allow applications to
respond to connection, disconnection, and IP acquisition events.
The API is modeled on a finite state machine with callbacks, allowing
applications to manage any socket connections in response to changes
in network connectivity.  Documentation and an example are provided.
A sub implemented is provided for the UNIX platform, for testing.