Remove use of CContext in port drivers.
This change set makes a broad set of changes across the network and socket drivers, replacing uses of CContext with the more simple Context structure.

The use of CContext in this code adds unnecessary complexity to the code, while proving no tangible benefit.  All of the gains are via preventing the GC from running inside the execution context of the port.

These changes are made under the terms of the LGPLv2 and Apache2 licenses.