Adds support for select (for debugging only)
In some cases it is nice to be able to debug on UNIX using select for checking the status of file descriptors (e.g., sockets).  This change adds support for select, with the addition of a compile time macro (defaults to using poll, for efficiency).  Select added to ESP32 port.

Also fixes two bugs:

* If the handler on an EventListener removes a listener from the global list of listeners, this has a side effect on the global->listeners pointer, so the check to exit the loop calling handlers may need to have the listeners value updated.  Without this change, we can end up in an infinite loop in some cases.

* Added the stdout to the calls to term_display that had been hidden in the ADAVANCED_TRACING macro.