Add support for some cryptographic NIFs.
This change set adds platform-specific NIFs for cryptographic operations, including:

* MD5 hashing
* random number generation
* atoms uniquely representing the platform:
	- generic_unix
	- eps32
	- stm32

Random number generation APIs that used to be ESP-specific are moved into the atomvm module,
as they are now supported on generic_unix platform, as well.

This change will attempt to find OpenSSL headers and libraries on the generic_unix platform.  If they are not found,
the associated NIFs will not be defined (similarly on platforms where the operations are not defined).