Add workaround to keep existing io module code working
AtomVM doesn't require a group leader. When is not used an invalid pid
is set (which is pid for process 0).
When invalid pid is set, erlang:group_leader returns self().
io functions assume that when group_leader is set to self(), a fallback
output nif should be used instead of IO protocol.