Add printf mnemonic defines for term hex/int representation
term is defined as uintptr_t, so any generic format placeholder will lead to
warnings, which is a blocking issue when -Werror is enforced.

stddint.h defines PRIuPTR and PRIXPTR, but they are not easy to
remember, and they expose term type implementation detail.
So add TERM_U_FMT and TERM_X_FMT.

Signed-off-by: Davide Bettio <davide@uninstall.it>