Update esp32 image tools
The bootloader offset is different for the ESP32-C3 and S3, so the release
tools now need to be generated depending on the target chip being built. The
`mkimage.sh` and `flashimage.sh` are no longer built with the generic_unix port,
instead they are created when the esp32 port is built so that the tool always
works with the last build.  The generated scripts are places in the esp32/build
folder, and can be used as `build/mkimage.sh` then `build/flashimage.sh` after
running `idf.py build`.

Signed-off-by: Winford <dwinford@pm.me>