Optimize memory usage by using memory heap fragments
Provide a new alloc mode to `memory_ensure_free_opt` when GC is not possible.

Simplify API to create a temporary heap on the stack.

Update API so term_* functions that allocate now take a Heap *. Update
signatures to follow guidelines in #535 (GlobalContext last, Heap penultimate).

Update several of these functions to no longer ensure free memory.

Optimize message passing by avoiding copying message terms twice. Message terms
are manipulated by the receiving process directly in the message's storage
and are garbage collected when the message is removed from the mailbox and
converted as a heap fragment of the receiving process.

Update documentation on heap fragments

Signed-off-by: Paul Guyot <pguyot@kallisys.net>