Add support for FP opcodes
These opcodes are officially supported since R8. `fclearerror/0` and
`fcheckerror/1` are noops as they are no longer implemented in OTP24 and the
error is always checked by other OP_F* opcodes.

No `badarith` exception is raised on underflow divisions, to be compatible
with BEAM (which cannot check that as it uses `isfinite(3)`). The only checked
exceptions are overflow and division by zero.

The tests were updated to actually generate the FP opcodes with OTP21-24
compilers. A new test was added for div.

Signed-off-by: Paul Guyot <pguyot@kallisys.net>