Improvements to stm32 build artifacts
This changes the location of the generated linker script to the `build`
directory, from the stm32 root directory. The `${DEVICE}` name is added
to the binary files. Binaries are in the root of the `build` directory,
rather than in `build/src`. Adds `-O2` compiler flag to reduce the size
of the final binary and speed up execution.

Signed-off-by: Winford <winford@object.stream>