Add actual support for =:= and =/=
This change makes possible to distinguish between 5.0 and 5 using =:= erlang
operator.

It has been implemented by adding an opts parameter to term_compare function,
which now has a `TermCompareExact` flag for exact term comparison.
TermCompareResult enum has not been changed and it still has the same
semantic.

There are still some TODOs about sorting operations (and similar) used
in maps that need to be addressed, and in general map comparison semantic is
not yet 100% compliant with OTP.

Signed-off-by: Davide Bettio <davide@uninstall.it>