Fix logic of `exit_reason` memory handling.
This is temporary fix as opposed to what is implemented in feature/smp.

`exit_reason` should be allocated in the target process. If it is allocated
in the ctx process, it will be copied to the target process during its next
garbage collection, however this will impact the used memory size, which is
used to ensure a given number of free terms. If `MIN_FREE_SPACE_SIZE` is lower
than the size of the exit reason, the GC will allocate less bytes than expected.

Signed-off-by: Paul Guyot <pguyot@kallisys.net>