Documentation build improvements
To help unify the theme and portability of the documentation the edoc generated
API documentation for the Erlang libraries is now created into markdown.  The
generated markdown is then built by Sphinx using the read-the-docs theme along
with the rest of the existing markdown documentation files from `doc/src`.
Logos and icons added to doc/src/_static.

Signed-off-by: Winford <winford@object.stream>