Unify documentation theme
Changes the way doxgen documention in rendered for src/libatomvm.  The final
html output is created by the `sphinx` `breathe` extension so it can be rendered
using the `read-the-docs` theme used be the rest of the documentation.

Note: there is no automation yet for adding listings to the catagorized sections
(including: functions, macros, types, and data structures), but the index and
content for the individual source files will be generated automatically. The
original `Doxygen` theme is still included and will always remain up to date.

Signed-off-by: Winford <winford@object.stream>