Fix a bug where native ports would be executed with no message
When a process is signaled, it is executed to process the signal, but the
mailbox may be empty after the signal is processed. This would yield a crash
on native handlers that expect `mailbox_first` to always return the message
to consume.

Also work around native ports that only process one message by rescheduling
them if there is another message.

Signed-off-by: Paul Guyot <pguyot@kallisys.net>