Fix a bug in ESP32 TCP client example (fix #592)
The example crashed on feature/smp branch (this is #592) because it called
`inet:peername/1` on a closed socket and received `{error, einval}` and did
not expect this return value. Yet this is the correct behavior.

Fix the example by calling `inet:peername/1` only once.

Also change the format strings so the packet will be printed properly once
bug #594 is fixed.

Signed-off-by: Paul Guyot <pguyot@kallisys.net>