Implement resource objects
Resources are garbage-collected memory-managed pointers that can have
destructors that are called when the pointer is garbage collected.
They are implemented as special refc binaries that look empty. This is similar
to BEAM's behavior before ERTS 9.0 (OTP-20.0).

Also create a header for enif-compatible API, to be used with resources. Add
`ERL_NIF_TERM` which is typedef'd to term, and update CodeQL rule accordingly.

Also reorder fields in Context structure so a Context can be passed as an enif
environment which could also be not linked to a context.

Signed-off-by: Paul Guyot <pguyot@kallisys.net>