Implement basic posix file functions as nifs
NIFs are compiled if platform features common POSIX functions `open`, `close`,
`read`, `write` and `unlink`.

Implementation is non-blocking and a more complete implementation of a file
module may require `enif_select`, at least to work with FIFO pipes and devices.

On ESP32, implementation depends on vfs_fat driver which is always blocking
(and does not support select). Qemu-based test writes to a (virtual) SD card.
However, support for SD card still requires C code to mount it which is not
the object of this PR.

Also fix detection of functions at link time with STM32 platform, even if
`open`, `close`, `read`, `write` and `unlink` can be linked with (but fail).

Also introduce utility function `posix_errno_to_term` that can be used to
convert an errno into an atom for known err numbers.

Signed-off-by: Paul Guyot <pguyot@kallisys.net>