Implement `enif_select` and add `select` and `mkfifo` to file functions
`enif_select` is an API that allows nifs to use `select(2)` or equivalent,
paving the way for nif-based drivers.

`enif_select` depends on platform implementations and this PR only implements
it on generic_unix platform, using `sys_poll_event` select-like function,
namely `kqueue(2)` or `poll(2)`. As a result, `enif_select` does not exactly
behaves like `select(2)` and this is documented.

For testing, add `mkfifo` and `select` to POSIX file functions.

Signed-off-by: Paul Guyot <pguyot@kallisys.net>