Update esp32 `flash.sh` tool for IDF 5.x compatibility
The esptool.py resides in a different in ESP-IDF v5.x than it does in v4.x, and
it can also be installed independantly from ESP-IDF. This updates the `flash.sh`
tool to use whichever esptool.py is active in the users path. This is prepended
to the `PATH` when the IDF is activated using `. ./export.sh`, so it will work
with both major versions. This also allows the user to flash the most recent
build again without having to first activate the ESP-IDF, if they have
esptool.py installed independantly, which is quite likely as it simplifies
working with rebar3.

Signed-off-by: Winford <winford@object.stream>