Fix issue with http_server introduced in 474dea9
Close TCP connection from http_server when reply/3 is used (as it used to do).
Connection is not closed by users of reply/4 as they may pass a Content-Length
header and let the client close the connection. This actually is a workaround
for an issue in the gen_tcp implementation where the socket is closed before
all data is sent.

Signed-off-by: Paul Guyot <pguyot@kallisys.net>