Improve published html documentation
Adds styled admonitions to the published documentation that highlight the
intention of "Notes" in the documentation. Using styles we can better draw
attention to important or critical warnings. These styled admonitions are a
common feature seen in the documentation of many projects, including ESP-IDF
and the new `ex_doc` versons of some of the core OTP modules.

Adds stlyed code blocks for better presentation of various Erlang, C, and
shell code examples in the published documentation.

Signed-off-by: Winford <winford@object.stream>