Allow locking pins used by stm32 system peripherals
To prevent accidental reset, or configuration of pins currently in use by system
peripherals on the stm32 platform, reserved pins (such as the CONSOLE) may be
locked using `sys_lock_pin`, and unlocked with `sys_unlock_pin` so they are not
unintentionally reset later when other peripherals such as gpio are configured.
This is also intended to support other drivers that consume gpio pins, such as
SPI and I2C, to avoid inadvertent resets or re-configuration by other drivers.
The libopencm3 libraries on the stm32 platform provide no protection from
re-confuguring a gpio pin that is in active use.

Signed-off-by: Winford <winford@object.stream>