Add AP mode support for Pico-W
Add the ability to configure the Pico-W wireless chipset in AP mode.
There is no DHCP server yet and stations cannot obtain an IP.
Also it is possible to configure the network driver in AP+STA mode, however
this does not work. It is still unclear whether this is a limitation of the
chipset.

Signed-off-by: Paul Guyot <pguyot@kallisys.net>