Fixes to installation.
This change set makes several changes to the installation of AtomVM on the generic_unix platform.

* A new CMake flag AVM_BUILD_RUNTIME_ONLY has been added (defaults to off), which when
enabled will tell the build to only build the AtomVM binary and libraries.
This can significantly shorten build times for systems that build AtomVM (e.g., Homebrew),
as it excludes the builds of tests and examples.

* The PackBEAM tool is no longer installed. The PackBEAM tool is not well suited
towards end-users, and is generally used as an internal build too.
Users should be directed to use the rebar3 and mix plugins for development work.

* The atomvmlib.avm is no longer renamed by the installer, for clarity and better tracking.

Signed-off-by: Fred Dushin <fred@dushin.net>