Enable stm32 core coupled memory
Enable using the Core Coupled Memory present in all stm32f4 and other mid to high
end stm32 MCUs. The 64k CCM is used for `.data` and `.bss` segments (aprox 2.5k)
as well as the `stack` for the VM. This leaves more of the 256k of slower sram
available for user application stack and heap. Interrupt handler code may also be
placed in the core coupled memory, this will eliminate delays from loading the
interupt handler function from flash.

Signed-off-by: Winford <winford@object.stream>