Remove blocking calls from stm32 gpio interrupt handlers
Removes the TRACE calls from interrupt hanlers becase these are blocking
calls and should not be used in interrupt handlers. Extra diagnostic info
is added to the `isr_handler` (that maintains the context and relays the
interrupt information to the `gpio_interrupt_callback`) that can be used to
determint the same information as the removed TRACE calls.

Signed-off-by: Winford <winford@object.stream>