Fix logic of workflows on push
Tag filter in several workflows meant any tag that doesn't contain a /.
We probably wanted every tag.

And if we have a filter on a tag but no filter on branches, the workflow is
only executed when a tag matches, regardless of path filtering which are always
ignored on tags.

Ref: https://docs.github.com/en/actions/using-workflows/workflow-syntax-for-github-actions#example-including-branches-and-tags

Signed-off-by: Paul Guyot <pguyot@kallisys.net>