All users to use logging macros without instantiating the logger manager.
This change set allows users to use logging macros and functions in code, but for those calls not to fail with an error if the logger manager has not been instantiated.

Users must still instantiate the logger manager (or it is instantiated automatically as part of the kernel application, TBD) in order for logs to pass the initial allow phase, but all logs will be silently dropped until the EAVMLib logger manager is instaniated.

Signed-off-by: Fred Dushin <fred@dushin.net>