Configure STM32 system clock speed and flash size
Configures main system clock and flash size based on the target device.
The console uart may be configured by using the `-DAVM_CFG_CONSOLE=` option, or
for Nucleo devices the on board TTL->USB-COM can be configured with the
`-DBOARD=nucleo` cmake option. Previously these values were hard coded in the C
sources and needed to be changed manually if a device other than the default
target was used.

Closes #454

Signed-off-by: Winford <winford@object.stream>