utils: add helper for `void *` to function pointer conversion
GCC raise a warning when performing `void *` to function pointer
conversion when pedantic is used.
Rather than disabling pedantic, add a helper function for performing the
conversion.
The helper will also check function pointer and void * type sizes in order to
check if this kind of conversion is safe or not.

Signed-off-by: Davide Bettio <davide@uninstall.it>