Fix broken links by add missing files in docs/CMakeLists.txt
There are several broken links on the `atomvm.net/doc` site due to the links
to files being unreachable during the build, these files are copied into the
documentation build source directory so that the links to them function as
expected.

A post build step has been added for the generated markdown documentation
files for the erlang modules. Several of the links that `edown` generates are
not processed correctly by the `myst_parser` and lead to broken links in the
html. The added post-build `sed` commands fix these links so that myst can
can generate working links in the html pages.

Signed-off-by: Winford <winford@object.stream>