Resolve doxygen problems processing opcodesshitch.h
Doxygen has a difficult time processing src/libAtomVM/opcosswitch.h, leading to
many warnings when building the documentation, and resulting to several
rendering errors in the generated documentation. There is no doxygen comment
documentation in this file, so it is better to omit it when generating the
documentation, rather than need to polute it with doxygen directives like the
previously added: `#ifndef DOXYGEN_SKIP_SECTION` comments.

The `Doxyfile.in` has been modified to exclude libAtomVM/opcosswitch.h from
processing, allowing for the removal of the previously inserted doxgen
directives.

Signed-off-by: Winford <winford@object.stream>