Fix rendering problems in publised html and pdf documentation
Adds pdf template and stylesheet to fix the way links to sections are rendered
by the spinx `rinoh` plugin, using the given section name for the link rather
than replacing them with the section heading number.

Adds custom.css to change the style of tables so they dont overflow display
and require side-scolling in the published html on atomvm.net.

Signed-off-by: Winford <winford@object.stream>