Fix specs for documented erlang functions
Fixes missing or incomplete specs for the published erlang libraries
documentation. For the parameters of a function to be incuded in the published
edocs they must have a matching parameter name along with the type
(Name :: type()) in the function spec, otherwise they will be excluded from the
published documentation and the user will not see the explanations of the
`@param` values.

Signed-off-by: Winford <winford@object.stream>