Add extra visual content to atomvm.net docs
The "AtomVM Internals" chapter has been enhanced, by including doxygen generated
content for some of the data structures and functions discussed in the guide.
This extra material is not included in the epub and pdf versions of the docs,
due to the fact that the rest of the doxygen C API documentation is already
excluded.

Signed-off-by: Winford <winford@object.stream>