Improve documentation links
Adds links to functions and modules documentation where they are discussed in
the various chapters and guides.  Changes some existing links to other guides
to more specific sections of the guide, where apropriate.

Some section names have been changed from repeated, or generic names to names
more specific to the section. This is done both to reduce build warnings
(repeated secion names), but more importantly make it possible to link to these
sections in the future, both in the documentation itself, but also to share
these links in social media, forum, or issue posts.

Fixes a few reference links to work in the pdf and epub versions of the docs,
by making them relative. (i.e. add missing `./` before filenames)

Signed-off-by: Winford <winford@object.stream>