Adjust some list rendering in documentation
Changes the unordered list descibing a javascript promise transaction from an
unordered to an ordered list (incrimentally numbered), since this is describing
a specific series of events that occur in a specific order.

Fixes the indention of some sub-lists. Markdown sublist items should have 2
spaces indention for each new sub-level.

Signed-off-by: Winford <winford@object.stream>