Make rp2040 wait for usb serial timeout configurable
Adds Pico `cmake` configuration option `AVM_USB_WAIT_SECONDS` to set a timeout
to wait for USB serial connection before starting the application. This is set
to 20 seconds by default, and can be changed by supplying the cmake option:
    -DAVM_USB_WAIT_SECONDS={seconds-to-wait}

This setting is overridden if `AVM_WAIT_FOR_USB_CONNECT=on` is used, which will
wait forever for a USB console connection before starting the application.

This allows testing `blinky` and network applications without requiring opening
a console connection to the device, while still waiting a reasonable ammount of
time for the user to make a console connection to the serial to see the output
of applications like `hello_world`.

Signed-off-by: Winford <winford@object.stream>