Do not check ATOMIC_POINTER_LOCK_FREE == 2
Even if it is not performant we don't need having lock free atomics, it
is just needed that they are correctly working and supported.

So this check is going to be removed in order to support smp.h even on
plarforms, such as ESP32-C3 where ATOMIC_POINTER_LOCK_FREE is 1 (that
means that it might not be lock free).

For atomic support we just rely on stdatomics.h existance and _Atomic
keyword being recognized.

Signed-off-by: Davide Bettio <davide@uninstall.it>