Adds new keys to erlang:system_info/1 for ESP-IDF semantic version
Adds the keys `esp_idf_semantic`, `esp_idf_major`, `esp_idf_minor`, and
`esp_idf_patch` to erlang:system_info/1 to obtain the semantic version of the
ESP-IDF that the current VM was build with.  This is intended to make it easier
for device drivers to know which ESP-IDF API versions to use when there are
breaking changes between ESP-IDF versions.

Signed-off-by: Winford <winford@object.stream>