Adds rp2040 support for os_ keys to erlang:system_info/1
Adds support for `erlang:system_info/1` keys `os_type` and `os_version` for
obtaining a conveniently parse-able semantic version of pico-sdk the VM is built
with. Adding `os_type`, which currently will return `{rp2040, pico-sdk}` is
mainly for completeness, but could be helpful to differentiate, if future ports
end up supporting the rp2040 platform, such as zephyr or nuttx.

Signed-off-by: Winford <winford@object.stream>