Add network:wifi_scan/0,1 to esp32 network driver
Adds the ability for devices configured for sta or ap+sta to scan for available access points. With
no arguments the default maximum results returned is 6. All default options can be configured in the
`sta_config()` section of the configuration used to start the network driver, or set in the
configuration parameter of `network:wifi_scan/1`.

Signed-off-by: Winford <winford@object.stream>