Completly stop driver and free all resources with network:stop/0 on ESP32
Adds a destroy callback to the ESP32 network driver to completely stop the
driver and free all network resources when network:stop/0 is used.  Previosly
the driver was not being stopped internally and resources were not freed when
the gen_server was stopped, causing instability, and possible crashes when
event callbacks were triggered, but there was no process alive to handle them.

Closes #643

Signed-off-by: Winford <winford@object.stream>