Have `idf.py build` assemble the complete ESP32 image
Rather than having the user or CI scripts run the extra ./build/mkimage.sh
sript, it will be run automatically if the `esp32boot.avm` libraries are found
in the `AtomVM/build/libs/esp32boot` directory from a previous generic_unix
build. This includes a prominet message for the user with instruction to flash
the complete image to the device.

Updates the esp32-build.yaml workflow to use the updated build steps, and only
create the necessary esp32 related libraries, excluding the creation of uf2 files.

Signed-off-by: Winford <winford@object.stream>